Splitted DBManagertests
